pyproject.toml
[project]
name = "india-first-mcp"
version = "0.1.0"
description = "India-First Multilingual Memory + Safety MCP for Autonomous Agents"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "langdetect>=1.0.9",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy[asyncio]>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[tool.setuptools.packages.find]
include = ["app*", "mcp*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
